STY: break 'wnd_manage_emission_lines.py' into multiple code files
