+++ pyxrf/gui_module/tests/test_wnd_manage_emission_lines.py
